Successful commit with category.

+++ petstore-services/src/main/java/com/iwd/petstore_services/PetStoreService.java
@@ -4,7 +4,7 @@ import com.iwd.petstore_services.domain.Pet;
 
 public interface PetStoreService {
 
-	public Pet get(Long petId);
+	public Pet get(Integer petId);
 
 	public Pet add(Pet petToBeInserted);
