+++ petstore-services/src/main/java/com/iwd/petstore_services/domain/Tag.java
@@ -4,24 +4,23 @@ import java.io.Serializable;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
 public class Tag implements Serializable {
 
 	@Id
-	@Column(name="TAG_ID")
-	private Long id;
+	@Column(name = "TAG_ID")
+	private Integer id;
 
-	@Column(name="TAG_NAME")
+	@Column(name = "TAG_NAME")
 	private String name;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -32,5 +31,5 @@ public class Tag implements Serializable {
 	public void setName(String name) {
 		this.name = name;
 	}
-	
+
 }
